+++ R/redcap_project.R
@@ -164,6 +164,8 @@ rc_make_datasets <- function(redcap_data) {
 
   }
 
+  class(datasets) <- c("odytools_datasets_list", class(datasets))
+
   datasets
 
 }
@@ -342,6 +344,11 @@ rc_back_up <- function() {
 #' @export
 ody_add_to_datasets <- function(object, description = NULL, export = FALSE) {
 
+  class(object) <- c("odytools_dataset", class(object))
+  # If the object is derived from an object already added to datasets, it will
+  # inherit the "odytools_dataset" class so this unique() ensures the class is
+  # not duplicated.
+  class(object) <- unique(class(object))
   attr(object, "is_dataset") <- TRUE
   attr(object, "description") <- description