+++ NEWS.md
@@ -1,7 +1,21 @@
+# odytools 0.9.0
+
+-   `ody_rc_import` now includes the attributes:
+    -   *atc_fields*: Vector of the field names coded with ATC.
+    -   *atc_codes*: A tibble with the ATC codes used by all *atc_fields*.
+-   `ody_rc_select` now adds the attributes *atc_fields* and *atc_codes* to the output in case any of the selected variables are coded with ATC.
+-   Added `ody_rc_translate_atc` to translate ATC codes to their descriptions.
+-   Datasets created inside an odytools REDCap project now belong to the S3 class *odytools_dataset* with an associated print method that shows the dataset's description. The datasets list itself also has an associated S3 class with a print method that lists all the datasets' and their descriptions.
+-   Added the argument `keep_user_na` to `ody_rc_format`. If TRUE, the user-defined missing values are kept as a character. If FALSE (the default), they are transformed into regular NA values.
+-   Added `ody_rc_get_metadata` to get the metadata of any REDCap project.
+-   Added `ody_add_tbl_violindotbox`. It adds a descriptive table below a plot created by `ody_plot_violindotplot`.
+-   Added `ody_apply_on_pattern` to apply a function to the columns of a data frame that match a given pattern.
+
 # odytools 0.8.1
+
 -   Improved the general appearance of some in-console messages.
 -   Added `ody_rc_add_label`. It labels the variables of a data frame according to the metadata of a REDCap project.
--   The resulting object of `ody_rc_import` now has the S3 class *odytools_redcap*. 
+-   The resulting object of `ody_rc_import` now has the S3 class *odytools_redcap*. This class has an associated print method that shows the project's name and the import date.
 -   Added `ody_plot_violindotplot` which plots a half violin, a jitter dot plot and boxplot in the same graph. It also performs comparisons between groups.
 
 # odytools 0.8.0